Simple uploads
A lot of work relating to file uploads. A few clientside classes have been added for handling uploads. One pertaining to the form data itself, which can have keys registered with it to wait. This is useful when you need to read the contents of a file in, or upload a file before submitting the form.

I've also added a RocketUploader class, which should take care of figuring out what kind of upload it should be. There are also some new helpers in Formtacular pertaining to these uploads.

It's also noteworthy that I've come up with a basic system to allow different kinds of upload. In total, there are (or will be) 3 ways someone can handle file uploads out of the box.

The first is the 'simple' method. This is somewhat like how a regular HTTP upload works. The contents of the file are read in their entirety, and dumped to a variable in the form. It works, but will be slow, and the upload won't be able to give feedback on upload progress. This is probably most advisable for very small files where the programmer doesn't want to deal with the headache of setting up one of the more powerful methods.

The second method is 'file_first', which delays sending the form data until the file has been uploaded. The idea is that the file will be read and sent in chunks. The form data will be sent once the file uploads are complete, and the controller action will recieve a param including the files locations for further processing. This method is fully asynchronus, and will allow developers to show a progress bar. It is probably best for instances where you don't want a record to be created until AFTER all attached files have been uploaded.

The last method is 'form_first', which sends the contents of the form immediately, and uploads the file after. Once the form hits the server, it needs to return the ID of the record that wants to be attached to the files. The RocketUploader will then begin uploading the file, and once complete, fire a callback on the object in question. This method probably makes the most sense when you don't mind a record showing no attachments until after the files have become available. For instance, say you are creating a wiki entry (please don't build content driven sites like wikis with Rocket!). You probably don't mind the article showing up as not having any attachments for a few minutes while the files upload.